+++ packages/socket/CHANGELOG.md
@@ -3,6 +3,18 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.8.0](https://github.com/ETransferProject/etransfer-toolkit/compare/v1.7.1-alpha.1...v1.8.0) (2024-10-14)
+
+**Note:** Version bump only for package @etransfer/socket
+
+## [1.7.1-alpha.1](https://github.com/ETransferProject/etransfer-toolkit/compare/v1.7.1-alpha.0...v1.7.1-alpha.1) (2024-09-25)
+
+**Note:** Version bump only for package @etransfer/socket
+
+## [1.7.1-alpha.0](https://github.com/ETransferProject/etransfer-toolkit/compare/v1.7.0...v1.7.1-alpha.0) (2024-09-25)
+
+**Note:** Version bump only for package @etransfer/socket
+
 ## [1.7.0](https://github.com/ETransferProject/etransfer-toolkit/compare/v1.6.5-alpha.3...v1.7.0) (2024-09-24)
 
 **Note:** Version bump only for package @etransfer/socket